Api key feature (#5)
This request features a basic user interface for registering a user with name, email and a domain for where the api is used. It also contains some queries for saving a user and key with it's owner ( a user ) and some endpoints for registering a user and checking if a given key is valid with hmac. For testing purposes the current hmac secret is not hidden.